requirements.txt
pybit>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
optuna>=3.4.0
websockets>=11.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
ta>=0.10.0
pandas-ta>=0.3.14
ccxt>=4.0.0
tenacity>=8.0.0
